refactor(application)  add whitespaces between properties
refactor(application)  add whitespaces between properties


refactor(core) fix class name, add whitespace


refactor(core) fix class name, add whitespace


refactor(application)  add whitespaces between properties
