+++ YesterdayApi/Core/Users/Followers/Follower.cs
@@ -3,6 +3,7 @@ namespace YesterdayApi.Core.Users.Followers
   public class Follower
   {
     public int FollowingUserId { get; set; }
+
     public int FollowedUserId { get; set; }
   }
 }
