+++ YesterdayApi/Core/Users/Posts/Comments/Web/CommentResponse.cs
@@ -5,11 +5,11 @@ namespace YesterdayApi.Core.Users.Posts.Comments.Web
   public class CommentResponse
   {
     public int Id { get; set; }
-    
+
     public string Content { get; set; }
-    
+
     public DateTime PublishDate { get; set; }
-    
+
     public int UserId { get; set; }
   }
 }
