+++ YesterdayApi/Core/Users/Posts/Post.cs
@@ -7,9 +7,13 @@ namespace YesterdayApi.Core.Users.Posts
   public class Post
   {
     public int Id { get; set; }
+
     public string Content { get; set; }
+
     public DateTime PublishDate { get; set; }
+
     public int UserId { get; set; }
+
     public ICollection<Reaction> Reactions { get; set; }
   }
 }
