+++ YesterdayApi/Core/Users/Posts/Reactions/Web/ReactionRequest.cs
@@ -5,8 +5,11 @@ namespace YesterdayApi.Core.Users.Posts.Reactions.Web
   public class ReactionRequest
   {
     public Type Type { get; set; }
+
     public DateTime PublishDate { get; set; }
+
     public int UserId { get; set; }
+
     public int PostId { get; set; }
   }
 }
