+++ YesterdayApi/Core/Users/Posts/Web/PostRequest.cs
@@ -3,6 +3,7 @@ namespace YesterdayApi.Core.Users.Posts.Web
   public class PostRequest
   {
     public string Content { get; set; }
+
     public int UserId { get; set; }
   }
 }
