+++ YesterdayApi/Core/Users/Posts/Web/PostResponse.cs
@@ -5,8 +5,11 @@ namespace YesterdayApi.Core.Users.Posts.Web
   public class PostResponse
   {
     public int Id { get; set; }
+
     public string Content { get; set; }
+
     public DateTime PublishDate { get; set; }
+
     public int UserId { get; set; }
   }
 }
