+++ YesterdayApi/Core/Users/User.cs
@@ -10,23 +10,23 @@ namespace YesterdayApi.Core.Users
   public class User
   {
     public int Id { get; set; }
-    
+
     public string Name { get; set; }
-    
+
     public string Surname { get; set; }
-    
+
     public string Login { get; set; }
-    
+
     public string Password { get; set; }
-    
+
     public DateTime BirthDate { get; set; }
-    
+
     public Profile Profile { get; set; }
 
     public UserRole UserRole { get; set; }
-    
+
     public ICollection<Reaction> Reactions { get; set; }
-    
+
     public ICollection<Comment> Comments { get; set; }
   }
 }
