+++ YesterdayApi/Core/Users/UserRoles/Roles/Role.cs
@@ -1,9 +1,13 @@
-namespace YesterdayApi.Core.Users.Roles.Role
+using YesterdayApi.Core.Users.Roles;
+
+namespace YesterdayApi.Core.Users.UserRoles.Roles
 {
   public class Role
   {
     public int Id { get; set; }
+
     public string Name { get; set; }
+
     public UserRole UserRole { get; set; }
   }
 }
