+++ YesterdayApi/Core/Users/UserRoles/UserRole.cs
@@ -3,6 +3,7 @@ namespace YesterdayApi.Core.Users.Roles
   public class UserRole
   {
     public int UserId { get; set; }
+
     public int RoleId { get; set; }
   }
 }
