+++ YesterdayApi/Core/Users/Web/UserRequest.cs
@@ -2,16 +2,16 @@ using System;
 
 namespace YesterdayApi.Core.Users.Web
 {
-  public class Request
+  public class UserRequest
   {
     public string Name { get; set; }
-    
+
     public string Surname { get; set; }
-    
+
     public string Login { get; set; }
-    
+
     public string Password { get; set; }
-    
+
     public DateTime BirthDate { get; set; }
   }
 }
